feat: refactor error handling by renaming validators to exceptions and updating imports
