+++ infrastructure/entry-points/reactive-web/src/main/java/co/com/pragma/api/exceptions/RequestValidator.java
@@ -1,6 +1,5 @@
-package co.com.pragma.api.validators;
+package co.com.pragma.api.exceptions;
 
-import co.com.pragma.api.exceptions.JakartaValidationException;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.Validator;
 import lombok.RequiredArgsConstructor;
