+++ infrastructure/entry-points/reactive-web/src/test/java/co/com/pragma/api/UserRouterRestTest.java
@@ -3,15 +3,13 @@ package co.com.pragma.api;
 import co.com.pragma.api.dto.request.CreateUserRequest;
 import co.com.pragma.api.dto.response.CreateUserResponse;
 import co.com.pragma.api.mappers.UserMapper;
-import co.com.pragma.api.validators.RequestValidator;
+import co.com.pragma.api.exceptions.RequestValidator;
 import co.com.pragma.model.user.User;
 import co.com.pragma.model.userrole.UserRole;
 import co.com.pragma.usecase.getuserroleusecase.GetUserRoleUseCase;
 import co.com.pragma.usecase.user.CreateUserUseCase;
-import lombok.RequiredArgsConstructor;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.reactive.WebFluxTest;
